MPauly Branch - Visible profile information
Note: I had to add a strategy to the generatedvalue, as well as update
database to utilize autoincrement values on uID in userprofile table.

Made user information visible on the edit profile page, as well as added
checks as to whether the user actually changed information upon clicking
save. 
Edited info will be displayed upon clicking edit profile again as well
as be updated in the database.